# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Avito.Kitchen — Kitchen Core API
  version: 0.2.0
  description: |
    Контракт Core для web-клиента и подключённых заведений. Авторизация исключена
    из scope тестового задания; в production все partner endpoint'ы требуют S2S auth.
servers:
  - url: http://localhost:8000
paths:
  /health:
    get:
      tags: [health]
      responses:
        '200': {description: Service is healthy}
  /client/establishments:
    get:
      tags: [client]
      summary: Список заведений, доступных для заказа
      responses:
        '200':
          description: OK
          content: {application/json: {schema: {$ref: '#/components/schemas/EstablishmentList'}}}
  /client/establishments/{establishment_id}/menu:
    get:
      tags: [client]
      summary: Меню заведения, включая недоступные позиции
      parameters: [{$ref: '#/components/parameters/EstablishmentId'}]
      responses:
        '200':
          description: OK
          content: {application/json: {schema: {$ref: '#/components/schemas/Menu'}}}
        '404': {$ref: '#/components/responses/NotFound'}
  /client/orders:
    post:
      tags: [client]
      summary: Создать заказ и зарезервировать остатки
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/OrderCreate'}}}
      responses:
        '201':
          description: Created
          content: {application/json: {schema: {$ref: '#/components/schemas/Order'}}}
        '409': {$ref: '#/components/responses/Conflict'}
    get:
      tags: [client]
      summary: История заказов пользователя
      parameters:
        - name: user_id
          in: query
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200':
          description: OK
          content: {application/json: {schema: {$ref: '#/components/schemas/OrderList'}}}
  /client/orders/{order_id}:
    get:
      tags: [client]
      summary: Статус и история заказа для polling
      parameters: [{$ref: '#/components/parameters/OrderId'}]
      responses:
        '200':
          description: OK
          content: {application/json: {schema: {$ref: '#/components/schemas/Order'}}}
        '404': {$ref: '#/components/responses/NotFound'}
  /client/orders/{order_id}/cancel:
    post:
      tags: [client]
      summary: Отменить заказ до решения заведения
      parameters: [{$ref: '#/components/parameters/OrderId'}]
      responses:
        '200':
          description: Cancelled
          content: {application/json: {schema: {$ref: '#/components/schemas/Order'}}}
        '409': {$ref: '#/components/responses/Conflict'}
  /establishments/{partner_id}:
    put:
      tags: [establishments]
      summary: Идемпотентно создать или обновить подключённое заведение
      parameters:
        - name: partner_id
          in: path
          required: true
          schema: {type: string, maxLength: 128}
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/EstablishmentUpsert'}}}
      responses:
        '200': {description: Updated, content: {application/json: {schema: {$ref: '#/components/schemas/Establishment'}}}}
        '201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/Establishment'}}}}
  /establishments/{establishment_id}/status:
    patch:
      tags: [establishments]
      summary: Открыть или закрыть заведение для новых заказов
      parameters: [{$ref: '#/components/parameters/EstablishmentId'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/EstablishmentStatusUpdate'}}}
      responses:
        '200': {description: Updated, content: {application/json: {schema: {$ref: '#/components/schemas/Establishment'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
  /establishments/{establishment_id}/categories:
    post:
      tags: [establishments]
      summary: Добавить категорию меню
      parameters: [{$ref: '#/components/parameters/EstablishmentId'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/CategoryCreate'}}}
      responses:
        '201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/Category'}}}}
        '409': {$ref: '#/components/responses/Conflict'}
  /establishments/{establishment_id}/products:
    post:
      tags: [establishments]
      summary: Создать позицию меню вручную
      parameters: [{$ref: '#/components/parameters/EstablishmentId'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ProductCreate'}}}
      responses:
        '201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/Product'}}}}
        '409': {$ref: '#/components/responses/Conflict'}
  /establishments/{establishment_id}/products/external/{external_id}:
    put:
      tags: [establishments]
      summary: Идемпотентная синхронизация позиции из POS
      parameters:
        - $ref: '#/components/parameters/EstablishmentId'
        - name: external_id
          in: path
          required: true
          schema: {type: string, maxLength: 128}
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ProductFields'}}}
      responses:
        '200': {description: Updated, content: {application/json: {schema: {$ref: '#/components/schemas/Product'}}}}
        '201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/Product'}}}}
  /establishments/{establishment_id}/products/{product_id}:
    patch:
      tags: [establishments]
      summary: Изменить доступность, остаток, цену или категорию
      parameters:
        - $ref: '#/components/parameters/EstablishmentId'
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/ProductUpdate'}}}
      responses:
        '200': {description: Updated, content: {application/json: {schema: {$ref: '#/components/schemas/Product'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
  /establishments/{establishment_id}/orders:
    get:
      tags: [establishments]
      summary: Очередь заказов заведения
      parameters:
        - $ref: '#/components/parameters/EstablishmentId'
        - name: status
          in: query
          schema: {$ref: '#/components/schemas/OrderStatus'}
      responses:
        '200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/OrderList'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
  /establishments/{establishment_id}/orders/{order_id}/decision:
    post:
      tags: [establishments]
      summary: Принять или отклонить заказ в статусе created
      parameters: [{$ref: '#/components/parameters/EstablishmentId'}, {$ref: '#/components/parameters/OrderId'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/OrderDecision'}}}
      responses:
        '200': {description: Updated, content: {application/json: {schema: {$ref: '#/components/schemas/Order'}}}}
        '409': {$ref: '#/components/responses/Conflict'}
  /establishments/{establishment_id}/orders/{order_id}/status:
    post:
      tags: [establishments]
      summary: Передвинуть заказ по автомату статусов
      parameters: [{$ref: '#/components/parameters/EstablishmentId'}, {$ref: '#/components/parameters/OrderId'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/OrderStatusUpdate'}}}
      responses:
        '200': {description: Updated, content: {application/json: {schema: {$ref: '#/components/schemas/Order'}}}}
        '409': {$ref: '#/components/responses/Conflict'}
components:
  parameters:
    EstablishmentId: {name: establishment_id, in: path, required: true, schema: {type: string, format: uuid}}
    OrderId: {name: order_id, in: path, required: true, schema: {type: string, format: uuid}}
    ProductId: {name: product_id, in: path, required: true, schema: {type: string, format: uuid}}
  responses:
    NotFound: {description: Resource not found}
    Conflict: {description: Business rule conflict}
  schemas:
    EstablishmentStatus: {type: string, enum: [open, closed]}
    OrderStatus: {type: string, enum: [created, accepted, rejected, preparing, ready, completed, cancelled]}
    Establishment:
      type: object
      required: [id, partner_id, name, status]
      properties:
        id: {type: string, format: uuid}
        partner_id: {type: string}
        name: {type: string}
        description: {type: [string, 'null']}
        status: {$ref: '#/components/schemas/EstablishmentStatus'}
    EstablishmentUpsert:
      type: object
      required: [name, callback_url]
      properties:
        name: {type: string, minLength: 1, maxLength: 255}
        description: {type: [string, 'null']}
        callback_url: {type: string, format: uri}
        status: {$ref: '#/components/schemas/EstablishmentStatus'}
    EstablishmentStatusUpdate:
      type: object
      required: [status]
      properties: {status: {$ref: '#/components/schemas/EstablishmentStatus'}}
    CategoryCreate:
      type: object
      required: [name]
      properties: {name: {type: string, minLength: 1, maxLength: 255}}
    Category:
      type: object
      required: [id, name]
      properties: {id: {type: string, format: uuid}, name: {type: string}}
    ProductFields:
      type: object
      required: [name, price]
      properties:
        category_id: {type: [string, 'null'], format: uuid}
        name: {type: string}
        description: {type: [string, 'null']}
        price: {type: string, pattern: '^\\d+\\.\\d{2}$', examples: ['590.00']}
        stock: {type: [integer, 'null'], minimum: 0}
        is_available: {type: boolean, default: true}
    ProductCreate:
      allOf:
        - $ref: '#/components/schemas/ProductFields'
        - type: object
          properties: {external_id: {type: [string, 'null'], maxLength: 128}}
    ProductUpdate:
      type: object
      minProperties: 1
      properties:
        category_id: {type: [string, 'null'], format: uuid}
        name: {type: string}
        description: {type: [string, 'null']}
        price: {type: string, pattern: '^\\d+\\.\\d{2}$'}
        stock: {type: [integer, 'null'], minimum: 0}
        is_available: {type: boolean}
    Product:
      allOf:
        - $ref: '#/components/schemas/ProductFields'
        - type: object
          required: [id]
          properties:
            id: {type: string, format: uuid}
            external_id: {type: [string, 'null']}
    Menu:
      type: object
      properties:
        establishment: {$ref: '#/components/schemas/Establishment'}
        categories: {type: array, items: {$ref: '#/components/schemas/Category'}}
        products: {type: array, items: {$ref: '#/components/schemas/Product'}}
    OrderItemIn:
      type: object
      required: [product_id, quantity]
      properties:
        product_id: {type: string, format: uuid}
        quantity: {type: integer, minimum: 1, maximum: 100}
    OrderCreate:
      type: object
      required: [user_id, establishment_id, delivery_address, items]
      properties:
        user_id: {type: string, format: uuid}
        establishment_id: {type: string, format: uuid}
        delivery_address: {type: string}
        comment: {type: [string, 'null']}
        items: {type: array, minItems: 1, maxItems: 50, items: {$ref: '#/components/schemas/OrderItemIn'}}
    OrderItem:
      type: object
      properties:
        product_id: {type: string, format: uuid}
        quantity: {type: integer}
        price_snapshot: {type: string, pattern: '^\\d+\\.\\d{2}$'}
    StatusHistory:
      type: object
      properties:
        status: {$ref: '#/components/schemas/OrderStatus'}
        changed_at: {type: string, format: date-time}
        comment: {type: [string, 'null']}
    Order:
      type: object
      properties:
        id: {type: string, format: uuid}
        user_id: {type: string, format: uuid}
        establishment_id: {type: string, format: uuid}
        delivery_address: {type: string}
        comment: {type: [string, 'null']}
        total_price: {type: string, pattern: '^\\d+\\.\\d{2}$'}
        status: {$ref: '#/components/schemas/OrderStatus'}
        rejection_reason: {type: [string, 'null']}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        items: {type: array, items: {$ref: '#/components/schemas/OrderItem'}}
        status_history: {type: array, items: {$ref: '#/components/schemas/StatusHistory'}}
    EstablishmentList: {type: array, items: {$ref: '#/components/schemas/Establishment'}}
    OrderList: {type: array, items: {$ref: '#/components/schemas/Order'}}